+++ packages/aws-cdk/lib/obsolete-flags.ts
@@ -1,0 +1,5 @@
+export const OBSOLETE_FLAGS = [
+  '@aws-cdk/core:enableStackNameDuplicates',
+  '@aws-cdk/aws-s3:grantWriteWithoutAcl',
+  '@aws-cdk/aws-kms:defaultKeyPolicies',
+];
